Implemented matrix multiplication and unit tested. Added Javadoc for project.

+++ src/mixedTerm.java
@@ -1,13 +1,25 @@
 import java.util.HashMap;
 
-
+/**
+ * A term containing multiply quantities, each raised to an exponent
+ * @author kylereeve
+ *
+ */
 public class mixedTerm {
-	private HashMap<Quantity,Integer> terms;
-	public mixedTerm(HashMap<Quantity, Integer> terms)
+	private HashMap<Quantity,Double> terms;
+	/**
+	 * constructor
+	 * @param terms Terms in the mixed term, each paired with a corresponding exponent
+	 */
+	public mixedTerm(HashMap<Quantity, Double> terms)
 	{
 		this.terms = terms;
 	}
-	public HashMap<Quantity, Integer> getTerms()
+	/**
+	 * Get the terms and associated exponents
+	 * @return
+	 */
+	public HashMap<Quantity, Double> getTerms()
 	{
 		return terms;
 	}
